move permissions to normal spot
