+++ api/base/permissions.py
@@ -11,7 +11,7 @@ from framework.auth.cas import CasResponse
 from osf.models import ApiOAuth2Application, ApiOAuth2PersonalToken
 from website.util.sanitize import is_iterable_but_not_string
 
-from osf.permissions import INSTITUTION_ADMIN
+from osf.utils.permissions import INSTITUTION_ADMIN
 
 # Implementation built on django-oauth-toolkit, but  with more granular control over read+write permissions
 #   https://github.com/evonove/django-oauth-toolkit/blob/d45431ea0bf64fd31e16f429db1e902dbf30e3a8/oauth2_provider/ext/rest_framework/permissions.py#L15
