--- osf/permissions.py
@@ -1,1 +1,0 @@
-INSTITUTION_ADMIN = 'institution_admin'
